# Conflicts:
#	VerticalWarehouses/Machine Automation Service/MAS_AutomationService/appsettings.Development.json
#	VerticalWarehouses/Machine Automation Service/MAS_DataLayer/DataLayer.cs
#	VerticalWarehouses/Machine Automation Service/Tests/MAS_DataLayerUnitTests/DataLayerUnitTest.cs

+++ Common/DataModels/DefaultCompartment.cs
@@ -19,9 +19,9 @@
 
         public string Note { get; set; }
 
-        public int XPosition { get; set; }
+        public double XPosition { get; set; }
 
-        public int YPosition { get; set; }
+        public double YPosition { get; set; }
 
         #endregion
     }
