+++ WMS/Data/Core/Providers/CompartmentTypeProvider.cs
@@ -56,9 +56,9 @@ namespace Ferretto.WMS.Data.Core.Providers
                     await this.dataContext.CompartmentTypes
                         .SingleOrDefaultAsync(
                             ct =>
-                                (ct.Width == model.Width && ct.Height == model.Height)
+                                ((int)ct.Width == (int)model.Width && (int)ct.Height == (int)model.Height)
                                 ||
-                                (ct.Width == model.Height && ct.Height == model.Width));
+                                ((int)ct.Width == (int)model.Height && (int)ct.Height == (int)model.Width));
 
                 if (existingCompartmentType == null)
                 {
